+++ src/starkbank/user.clj
@@ -42,7 +42,7 @@
   each request or may be defined as the default user at the start (See README).
   If you are accessing a specific Workspace using Organization credentials, you should
   specify the workspace ID when building the Organization map or by request, using
-  the Organization.replace(organization, workspace_id) method, which creates a copy of the organization
+  the user/organization-replace(organization, workspace_id) method, which creates a copy of the organization
   map with the altered workspace ID. If you are listing or creating new Workspaces, the
   workspace_id should be nil.
 
@@ -82,3 +82,7 @@
    (try
      (get-java-user params)
      (catch Exception e (callback params)))))
+
+(defn organization-replace
+  ([organization, workspace-id]
+    (update organization :workspace-id #(str workspace-id %))))