+++ test/starkbank/brcode_preview_test.clj
@@ -6,7 +6,7 @@
   
 (deftest query-brcode-preview
 	(testing "query brcode-preview"
-		(user/set-test-user)
+		(user/set-test-project)
 		(def invoices (invoice/query {:limit 1 :status "created"}))
 		(def brcodes [(:brcode (first invoices))])
 		(def preview (first (preview/query {:brcodes brcodes})))